# Fix a bug and run the optimisation with nonzero ascending nodes