--- a/astronomy/astronomy.vcxproj.filters
+++ b/astronomy/astronomy.vcxproj.filters
@@ -100,6 +100,9 @@
     <Text Include="trappist_gravity_model.proto.txt">
       <Filter>Resource Files</Filter>
     </Text>
+    <Text Include="trappist_preoptimisation_initial_state_jd_2457000_000000000.proto.txt">
+      <Filter>Resource Files</Filter>
+    </Text>
     <Text Include="trappist_initial_state_jd_2457000_000000000.proto.txt">
       <Filter>Resource Files</Filter>
     </Text>