… astronomy/trappist_preoptimisation_initial_state_jd_2457000_000000000.proto.txt